Refactored view directory to views and moved static directory to resources.
