+++ src/main/kotlin/com/badger/taskschedulingapp/Main/MyApp.kt
@@ -1,7 +1,7 @@
 package com.badger.taskschedulingapp.Main
 
 import com.badger.taskschedulingapp.Main.staic.Styles
-import com.badger.taskschedulingapp.Main.view.MainView
+import com.badger.taskschedulingapp.Main.views.MainView
 import tornadofx.*
 
 class MyApp: App(MainView::class, Styles::class){
