+++ src/main/resources/staic/Styles.kt
